+++ src/main/java/com/caiata/utils/GlobalParameters.java
@@ -1,4 +1,4 @@
-package com.caiata;
+package com.caiata.utils;
 
 import java.io.File;