+++ src/test/java/com/caiata/web/TestEbay.java
@@ -1,14 +1,17 @@
 package com.caiata.web;
 
 import com.caiata.steps.EbaySteps;
-import com.caiata.ManagementDriver;
-import com.caiata.Utility;
+import com.caiata.utils.ManagementDriver;
+import com.caiata.utils.ModelloEbay;
+import com.caiata.utils.Utility;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+
+import java.util.ArrayList;
 import java.util.Properties;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -17,6 +20,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 public class TestEbay {
 
+    static private ModelloEbay modello = new ModelloEbay();
     static private WebDriver driver = null;
     static private Properties webProp = null;
     static private EbaySteps steps = null;
@@ -70,6 +74,31 @@ public class TestEbay {
         steps.closeBanner(webProp);
         steps.selezionaCategoria(webProp, categoria);
         assertTrue(driver.findElement(By.xpath(webProp.getProperty(key))).isEnabled());
+    }
+
+    @Order(4)
+    @Test
+    @DisplayName("Stampa delle categorie.")
+    void test_004(){
+        driver.get(webProp.getProperty("ebay.url"));
+        steps.closeBanner(webProp);
+        String[] tmp = steps.getMenuCategory(webProp);
+        System.out.println("----CATEGORIE----");
+        for(int i = 0; i < tmp.length ; i++){
+            System.out.println(tmp[i]);
+        }
+    }
+
+    @Order(5)
+    @ParameterizedTest(name = "q = {0} , categoria = {0}")
+    @CsvSource({"iphone, cell phone"})
+    @DisplayName("Stampa di tutti i risultati.")
+    void test_005(String q, String categoria) {
+        driver.get(webProp.getProperty("ebay.url"));
+        steps.closeBanner(webProp);
+        steps.selezionaCategoria(webProp, categoria);
+        steps.search(webProp, q);
+        modello.stampaElementi(steps.getElementi(webProp));
     }
 
     @AfterEach