+++ CHANGELOG.md
@@ -13,3 +13,4 @@ _`**VERSIONE 1.0**`_
 - Aggiunto `Test per le categorie Ebay`
 - Aggiunto `Test per la selezione delle categorie Ebay`
 - Aggiunti `fluent wait e webDriverWait`
+- Aggiunto `ModelloEbay`